+++ src/service/index.ts
@@ -1,0 +1,13 @@
+import { IGlobalService } from './GlobalService';
+import { IPageService } from './PageService';
+import { IComponentService } from './ComponentService';
+import { IEntityService } from './EntityService';
+import { IDataRecordService } from './DataRecordService';
+
+export {
+  IGlobalService,
+  IPageService,
+  IComponentService,
+  IEntityService,
+  IDataRecordService,
+}