+++ src/types/index.ts
@@ -1,10 +1,9 @@
-import React from 'react';
+import React, { Component } from 'react';
 import EventEmitter from 'eventemitter3';
 import { AxiosInstance } from 'axios';
 import { createBEM } from '../bem';
 import { loadUmdLib, loadCss } from '../load';
 import { transMetaVal2PureVal, transObjectKeys } from '../trans';
-import { useDynamicValue } from '../use-dynamic-value';
 import {
   CanvasContextData,
   ContainerContextData,
@@ -24,21 +23,36 @@ import { UseEventHandlers } from './use-event-handlers';
 import { DesignModeBuiltIns, RunModeBuiltIns } from './builtins';
 import { UseDataSourceApi, UsePropertiesApi } from './properties';
 import { UseCanvasApi } from './canvas';
-import { UseComponentStateApi, UseComponentSubscriberApi, UseModelApi } from './observer';
 import { UseEventBusApi } from './eventbus';
+import { UseDynamicValueApi } from './dynamic-value';
+import {
+  UseComponentServiceApi,
+  UsePageServiceApi,
+  UseGlobalServiceApi,
+  UseEntityServiceApi,
+  UseDataRecordServiceApi,
+  UseComponentStateApi,
+  UseComponentActionApi,
+  UseComponentSubscriberApi,
+  UseModelApi,
+  UseExpressionApi,
+  UseExpressionWithContextApi,
+  UseVariableApi,
+  UseParserApi,
+} from '../hooks';
+import { IGlobalService, IPageService, IComponentService, IDataRecordService, IEntityService } from '../service';
 
 export * from './component-key';
 export * from './container-context-data';
 export * from './dynamic-value';
 export * from './fetch';
 export * from './jsx-component';
-export * from './panel-config';
+export * from '../components/properties';
 export * from './trans';
 export * from './use-event-handlers';
 export * from './model';
 export * from './properties';
 export * from './canvas';
-export * from './observer';
 
 export { DesignModeBuiltIns, RunModeBuiltIns };
 
@@ -58,7 +72,6 @@ export type CommonSDK = {
   loadCss: typeof loadCss;
   transMetaVal2PureVal: typeof transMetaVal2PureVal;
   transObjectKeys: typeof transObjectKeys;
-  useDynamicValue: typeof useDynamicValue;
   useCanvas: UseCanvasApi;
 };
 
@@ -73,16 +86,46 @@ export type DesignModeSDK = CommonSDK & DesignModeSDKInjection;
 export type RunModeSDKInjection = {
   useEventHandlers: UseEventHandlers;
   request: AxiosInstance;
+    /**
+   * @deprecated 已废弃，不建议通过SDK使用, 建议使用SDK.getEntityService('yourEntityApiKey').createRecord
+   */
   createRecord: CreateRecordMethod;
+  /**
+   * @deprecated 已废弃，不建议通过SDK使用, 建议使用SDK.getEntityService('yourEntityApiKey').deleteRecord
+   */
   deleteRecords: DeleteRecordsMethod;
+  /**
+  * @deprecated 已废弃，不建议通过SDK使用, 建议使用SDK.getEntityService('yourEntityApiKey').updateRecord
+  */
   modifyRecordAttribute: ModifyRecordAttributeMethod;
+  /**
+   * @deprecated 已废弃，不建议通过SDK使用, 建议使用SDK.getEntityService('yourEntityApiKey').getRecordByFilter
+   */
   fetchRecordsByCondition: FetchRecordsByConditionMethod;
+  /**
+   * @deprecated 已废弃，不建议通过SDK使用, 建议使用const { fetchData } = SDK.useDataSource(dataSource);
+   */
   fetchByDataSource: FetchByDataSourceMethod;
   useComponentState: UseComponentStateApi;
+  useComponentAction: UseComponentActionApi;
   useComponentSubscriber: UseComponentSubscriberApi;
   useModel: UseModelApi;
   useEventBus: UseEventBusApi;
   useDataSource: UseDataSourceApi;
+  useDynamicValue: UseDynamicValueApi;
+  useComponentService: UseComponentServiceApi;
+  usePageService: UsePageServiceApi;
+  useGlobalService: UseGlobalServiceApi;
+  useEntityService: UseEntityServiceApi;
+  useDataRecordService: UseDataRecordServiceApi;
+  useExpressionApi: UseExpressionApi,
+  useExpressionWithContextApi: UseExpressionWithContextApi,
+  useVariable: UseVariableApi,
+  useParser: UseParserApi,
+  getGlobalService: IGlobalService.getService,
+  getPageService: IPageService.getService,
+  getComponentService: IComponentService.getService,
+  getEntityService: IEntityService.getService,
 };
 
 export type RunModeSDK = CommonSDK & RunModeSDKInjection;